Initial integration testing implementation
