+++ server/db.js
@@ -1,10 +1,10 @@
 const MongoClient = require('mongodb').MongoClient
 const MongoMemoryServer = require('mongodb-memory-server').MongoMemoryServer
+
 const FAKE_DB = process.env.DB === 'fake'
 const MONGO_URL = process.env.MONGO || 'mongodb://localhost:27017/db'
-const fakeData = require('./fake_db.json')
-const loadDBwithData = require('./utils').loadDBwithData
 
+let mongod = null
 let client = null
 let db = null
 
@@ -18,7 +18,7 @@ module.exports = () => {
 module.exports.init = async () => {
     let mongo_url = MONGO_URL
     if (FAKE_DB) {
-        const mongod = new MongoMemoryServer()
+        mongod = new MongoMemoryServer()
         mongo_url = await mongod.getConnectionString()
         console.log('Using fake mongo at ' + mongo_url)
     } else {
@@ -30,9 +30,11 @@ module.exports.init = async () => {
     await client.connect()
     db = client.db('bob')
 
-    if (FAKE_DB) {
-        await loadDBwithData(db, fakeData)
-        console.log('Initialized database with fake data.')
-    }
     return db
+}
+
+module.exports.close = () => {
+    if (client) client.close()
+    //if (db) db.close()
+    if (mongod) mongod.stop()
 }
